phi4-modular/requirements.txt
numpy>=1.20.0
requests>=2.25.0
Pillow>=9.0.0
tqdm>=4.60.0
datasets>=2.10.0
huggingface-hub>=0.12.0
pathlib>=1.0.1
rouge_score
bert_score
evaluate
pydantic
wandb